+++ src/rp2_common/pico_cyw43_arch/include/pico/cyw43_arch/arch_freertos.h
@@ -7,12 +7,6 @@
 #ifndef _PICO_CYW43_ARCH_ARCH_FREERTOS_H
 #define _PICO_CYW43_ARCH_ARCH_FREERTOS_H
 
-#include "pico/cyw43_arch/arch_common.h"
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
 // PICO_CONFIG: CYW43_TASK_STACK_SIZE, Stack size for the CYW43 FreeRTOS task in 4-byte words, type=int, default=1024, group=pico_cyw43_arch
 #ifndef CYW43_TASK_STACK_SIZE
 #define CYW43_TASK_STACK_SIZE 1024
@@ -21,10 +15,6 @@ extern "C" {
 // PICO_CONFIG: CYW43_TASK_PRIORITY, Priority for the CYW43 FreeRTOS task, type=int default=4, group=pico_cyw43_arch
 #ifndef CYW43_TASK_PRIORITY
 #define CYW43_TASK_PRIORITY (tskIDLE_PRIORITY + 4)
-#endif
-
-#ifdef __cplusplus
-}
 #endif
 
 #endif
