+++ src/rp2_common/pico_cyw43_arch/include/pico/cyw43_arch/arch_poll.h
@@ -7,16 +7,6 @@
 #ifndef _PICO_CYW43_ARCH_ARCH_POLL_H
 #define _PICO_CYW43_ARCH_ARCH_POLL_H
 
-#include "pico/cyw43_arch/arch_common.h"
-
-#include <assert.h>
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-#ifdef __cplusplus
-}
-#endif
+// now obsolete; kept for backwards compatibility
 
 #endif
