+++ src/rp2_common/pico_cyw43_arch/include/pico/cyw43_arch/arch_threadsafe_background.h
@@ -7,14 +7,6 @@
 #ifndef _PICO_CYW43_ARCH_ARCH_THREADSAFE_BACKGROUND_H
 #define _PICO_CYW43_ARCH_ARCH_THREADSAFE_BACKGROUND_H
 
-#include "pico/cyw43_arch/arch_common.h"
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-#ifdef __cplusplus
-}
-#endif
+// now obsolete; kept for backwards compatibility
 
 #endif
